Fix pitch bend.
First, store the raw data within the event, so it's not so easy to try
to write a short instead of two bytes to the device.

Next, drop the offsetting for now. If that's necessary for printing
events, we can still subtract the "middle" value 0x200.

Last, make it clear that the value of pitch bend events comes with the
LSB first, for whatever reason.